Shell: Make 'editor' a member of Shell, and provide a LibShell
