+++ Shell/Builtin.cpp
@@ -34,7 +34,6 @@
 #include <unistd.h>
 
 extern char** environ;
-extern RefPtr<Line::Editor> editor;
 
 int Shell::builtin_alias(int argc, const char** argv)
 {
@@ -418,8 +417,8 @@ int Shell::builtin_disown(int argc, const char** argv)
 
 int Shell::builtin_history(int, const char**)
 {
-    for (size_t i = 0; i < editor->history().size(); ++i) {
-        printf("%6zu  %s\n", i, editor->history()[i].characters());
+    for (size_t i = 0; i < m_editor->history().size(); ++i) {
+        printf("%6zu  %s\n", i, m_editor->history()[i].characters());
     }
     return 0;
 }
