+++ Shell/CMakeLists.txt
@@ -6,8 +6,14 @@ set(SOURCES
     NodeVisitor.cpp
     Parser.cpp
     Shell.cpp
+)
+
+serenity_lib(LibShell shell)
+target_link_libraries(LibShell LibCore LibLine)
+
+set(SOURCES
     main.cpp
 )
 
 serenity_bin(Shell)
-target_link_libraries(Shell LibCore LibLine)
+target_link_libraries(Shell LibShell)
